@@ -1,23 +1,23 @@
 module github.com/dexidp/dex/examples
 
-go 1.21
+go 1.23
 
 require (
-	github.com/coreos/go-oidc/v3 v3.10.0
-	github.com/dexidp/dex/api/v2 v2.1.0
-	github.com/spf13/cobra v1.8.1
-	golang.org/x/oauth2 v0.20.0
-	google.golang.org/grpc v1.64.0
+	github.com/coreos/go-oidc/v3 v3.12.0
+	github.com/dexidp/dex/api/v2 v2.3.0
+	github.com/spf13/cobra v1.9.1
+	golang.org/x/oauth2 v0.26.0
+	google.golang.org/grpc v1.70.0
 )
 
 require (
-	github.com/go-jose/go-jose/v4 v4.0.1 // indirect
+	github.com/go-jose/go-jose/v4 v4.0.5 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/spf13/pflag v1.0.5 // indirect
-	golang.org/x/crypto v0.21.0 // indirect
-	golang.org/x/net v0.23.0 // indirect
-	golang.org/x/sys v0.18.0 // indirect
-	golang.org/x/text v0.14.0 // indirect
-	google.golang.org/genproto/googleapis/rpc v0.0.0-20240318140521-94a12d6c2237 // indirect
-	google.golang.org/protobuf v1.33.0 // indirect
+	github.com/spf13/pflag v1.0.6 // indirect
+	golang.org/x/crypto v0.32.0 // indirect
+	golang.org/x/net v0.33.0 // indirect
+	golang.org/x/sys v0.29.0 // indirect
+	golang.org/x/text v0.21.0 // indirect
+	google.golang.org/genproto/googleapis/rpc v0.0.0-20241202173237-19429a94021a // indirect
+	google.golang.org/protobuf v1.36.5 // indirect
 )
