@@ -1,22 +1,21 @@
 module github.com/dexidp/dex
 
-go 1.21
+go 1.24
 
 require (
-	cloud.google.com/go/compute/metadata v0.3.0
-	entgo.io/ent v0.13.1
-	github.com/AppsFlyer/go-sundheit v0.5.0
+	cloud.google.com/go/compute/metadata v0.6.0
+	entgo.io/ent v0.14.2
+	github.com/AppsFlyer/go-sundheit v0.6.0
 	github.com/Masterminds/semver v1.5.0
-	github.com/Masterminds/sprig/v3 v3.2.3
-	github.com/beevik/etree v1.4.0
-	github.com/coreos/go-oidc/v3 v3.10.0
-	github.com/dexidp/dex/api/v2 v2.1.0
-	github.com/felixge/httpsnoop v1.0.4
-	github.com/fsnotify/fsnotify v1.7.0
+	github.com/Masterminds/sprig/v3 v3.3.0
+	github.com/beevik/etree v1.5.0
+	github.com/coreos/go-oidc/v3 v3.12.0
+	github.com/dexidp/dex/api/v2 v2.3.0
+	github.com/fsnotify/fsnotify v1.8.0
 	github.com/ghodss/yaml v1.0.0
-	github.com/go-jose/go-jose/v4 v4.0.2
-	github.com/go-ldap/ldap/v3 v3.4.6
-	github.com/go-sql-driver/mysql v1.8.1
+	github.com/go-jose/go-jose/v4 v4.0.5
+	github.com/go-ldap/ldap/v3 v3.4.10
+	github.com/go-sql-driver/mysql v1.9.0
 	github.com/google/uuid v1.6.0
 	github.com/gorilla/handlers v1.5.2
 	github.com/gorilla/mux v1.8.1
