@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go. DO NOT EDIT.
 // versions:
-// 	protoc-gen-go v1.32.0
-// 	protoc        v4.24.4
+// 	protoc-gen-go v1.36.5
+// 	protoc        v5.29.3
 // source: server/internal/types.proto
 
 // Package internal holds protobuf types used by the server.
@@ -13,6 +13,7 @@ import (
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
 	reflect "reflect"
 	sync "sync"
+	unsafe "unsafe"
 )
 
 const (
@@ -24,21 +25,18 @@ const (
 
 // RefreshToken is a message that holds refresh token data used by dex.
 type RefreshToken struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	RefreshId     string                 `protobuf:"bytes,1,opt,name=refresh_id,json=refreshId,proto3" json:"refresh_id,omitempty"`
+	Token         string                 `protobuf:"bytes,2,opt,name=token,proto3" json:"token,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	RefreshId string `protobuf:"bytes,1,opt,name=refresh_id,json=refreshId,proto3" json:"refresh_id,omitempty"`
-	Token     string `protobuf:"bytes,2,opt,name=token,proto3" json:"token,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *RefreshToken) Reset() {
 	*x = RefreshToken{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_server_internal_types_proto_msgTypes[0]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_server_internal_types_proto_msgTypes[0]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *RefreshToken) String() string {
@@ -49,7 +47,7 @@ func (*RefreshToken) ProtoMessage() {}
 
 func (x *RefreshToken) ProtoReflect() protoreflect.Message {
 	mi := &file_server_internal_types_proto_msgTypes[0]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -81,21 +79,18 @@ func (x *RefreshToken) GetToken() string {
 // IDTokenSubject represents both the userID and connID which is returned
 // as the "sub" claim in the ID Token.
 type IDTokenSubject struct {
-	state         protoimpl.MessageState
-	sizeCache     protoimpl.SizeCache
+	state         protoimpl.MessageState `protogen:"open.v1"`
+	UserId        string                 `protobuf:"bytes,1,opt,name=user_id,json=userId,proto3" json:"user_id,omitempty"`
+	ConnId        string                 `protobuf:"bytes,2,opt,name=conn_id,json=connId,proto3" json:"conn_id,omitempty"`
 	unknownFields protoimpl.UnknownFields
-
-	UserId string `protobuf:"bytes,1,opt,name=user_id,json=userId,proto3" json:"user_id,omitempty"`
-	ConnId string `protobuf:"bytes,2,opt,name=conn_id,json=connId,proto3" json:"conn_id,omitempty"`
+	sizeCache     protoimpl.SizeCache
 }
 
 func (x *IDTokenSubject) Reset() {
 	*x = IDTokenSubject{}
-	if protoimpl.UnsafeEnabled {
-		mi := &file_server_internal_types_proto_msgTypes[1]
-		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
-		ms.StoreMessageInfo(mi)
-	}
+	mi := &file_server_internal_types_proto_msgTypes[1]
+	ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+	ms.StoreMessageInfo(mi)
 }
 
 func (x *IDTokenSubject) String() string {
@@ -106,7 +101,7 @@ func (*IDTokenSubject) ProtoMessage() {}
 
 func (x *IDTokenSubject) ProtoReflect() protoreflect.Message {
 	mi := &file_server_internal_types_proto_msgTypes[1]
-	if protoimpl.UnsafeEnabled && x != nil {
+	if x != nil {
 		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
 		if ms.LoadMessageInfo() == nil {
 			ms.StoreMessageInfo(mi)
@@ -137,7 +132,7 @@ func (x *IDTokenSubject) GetConnId() string {
 
 var File_server_internal_types_proto protoreflect.FileDescriptor
 
-var file_server_internal_types_proto_rawDesc = []byte{
+var file_server_internal_types_proto_rawDesc = string([]byte{
 	0x0a, 0x1b, 0x73, 0x65, 0x72, 0x76, 0x65, 0x72, 0x2f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61,
 	0x6c, 0x2f, 0x74, 0x79, 0x70, 0x65, 0x73, 0x2e, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x12, 0x08, 0x69,
 	0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x22, 0x43, 0x0a, 0x0c, 0x52, 0x65, 0x66, 0x72, 0x65,
@@ -153,22 +148,22 @@ var file_server_internal_types_proto_rawDesc = []byte{
 	0x65, 0x78, 0x69, 0x64, 0x70, 0x2f, 0x64, 0x65, 0x78, 0x2f, 0x73, 0x65, 0x72, 0x76, 0x65, 0x72,
 	0x2f, 0x69, 0x6e, 0x74, 0x65, 0x72, 0x6e, 0x61, 0x6c, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f,
 	0x33,
-}
+})
 
 var (
 	file_server_internal_types_proto_rawDescOnce sync.Once
-	file_server_internal_types_proto_rawDescData = file_server_internal_types_proto_rawDesc
+	file_server_internal_types_proto_rawDescData []byte
 )
 
 func file_server_internal_types_proto_rawDescGZIP() []byte {
 	file_server_internal_types_proto_rawDescOnce.Do(func() {
-		file_server_internal_types_proto_rawDescData = protoimpl.X.CompressGZIP(file_server_internal_types_proto_rawDescData)
+		file_server_internal_types_proto_rawDescData = protoimpl.X.CompressGZIP(unsafe.Slice(unsafe.StringData(file_server_internal_types_proto_rawDesc), len(file_server_internal_types_proto_rawDesc)))
 	})
 	return file_server_internal_types_proto_rawDescData
 }
 
 var file_server_internal_types_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
-var file_server_internal_types_proto_goTypes = []interface{}{
+var file_server_internal_types_proto_goTypes = []any{
 	(*RefreshToken)(nil),   // 0: internal.RefreshToken
 	(*IDTokenSubject)(nil), // 1: internal.IDTokenSubject
 }
@@ -185,37 +180,11 @@ func file_server_internal_types_proto_init() {
 	if File_server_internal_types_proto != nil {
 		return
 	}
-	if !protoimpl.UnsafeEnabled {
-		file_server_internal_types_proto_msgTypes[0].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*RefreshToken); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-		file_server_internal_types_proto_msgTypes[1].Exporter = func(v interface{}, i int) interface{} {
-			switch v := v.(*IDTokenSubject); i {
-			case 0:
-				return &v.state
-			case 1:
-				return &v.sizeCache
-			case 2:
-				return &v.unknownFields
-			default:
-				return nil
-			}
-		}
-	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
 		File: protoimpl.DescBuilder{
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
-			RawDescriptor: file_server_internal_types_proto_rawDesc,
+			RawDescriptor: unsafe.Slice(unsafe.StringData(file_server_internal_types_proto_rawDesc), len(file_server_internal_types_proto_rawDesc)),
 			NumEnums:      0,
 			NumMessages:   2,
 			NumExtensions: 0,
@@ -226,7 +195,6 @@ func file_server_internal_types_proto_init() {
 		MessageInfos:      file_server_internal_types_proto_msgTypes,
 	}.Build()
 	File_server_internal_types_proto = out.File
-	file_server_internal_types_proto_rawDesc = nil
 	file_server_internal_types_proto_goTypes = nil
 	file_server_internal_types_proto_depIdxs = nil
 }
