@@ -1,7 +1,7 @@
 // Code generated by protoc-gen-go-grpc. DO NOT EDIT.
 // versions:
-// - protoc-gen-go-grpc v1.3.0
-// - protoc             v4.24.4
+// - protoc-gen-go-grpc v1.5.1
+// - protoc             v5.29.3
 // source: api/v2/api.proto
 
 package api
@@ -15,27 +15,34 @@ import (
 
 // This is a compile-time assertion to ensure that this generated file
 // is compatible with the grpc package it is being compiled against.
-// Requires gRPC-Go v1.32.0 or later.
-const _ = grpc.SupportPackageIsVersion7
+// Requires gRPC-Go v1.64.0 or later.
+const _ = grpc.SupportPackageIsVersion9
 
 const (
-	Dex_GetClient_FullMethodName      = "/api.Dex/GetClient"
-	Dex_CreateClient_FullMethodName   = "/api.Dex/CreateClient"
-	Dex_UpdateClient_FullMethodName   = "/api.Dex/UpdateClient"
-	Dex_DeleteClient_FullMethodName   = "/api.Dex/DeleteClient"
-	Dex_CreatePassword_FullMethodName = "/api.Dex/CreatePassword"
-	Dex_UpdatePassword_FullMethodName = "/api.Dex/UpdatePassword"
-	Dex_DeletePassword_FullMethodName = "/api.Dex/DeletePassword"
-	Dex_ListPasswords_FullMethodName  = "/api.Dex/ListPasswords"
-	Dex_GetVersion_FullMethodName     = "/api.Dex/GetVersion"
-	Dex_ListRefresh_FullMethodName    = "/api.Dex/ListRefresh"
-	Dex_RevokeRefresh_FullMethodName  = "/api.Dex/RevokeRefresh"
-	Dex_VerifyPassword_FullMethodName = "/api.Dex/VerifyPassword"
+	Dex_GetClient_FullMethodName       = "/api.Dex/GetClient"
+	Dex_CreateClient_FullMethodName    = "/api.Dex/CreateClient"
+	Dex_UpdateClient_FullMethodName    = "/api.Dex/UpdateClient"
+	Dex_DeleteClient_FullMethodName    = "/api.Dex/DeleteClient"
+	Dex_CreatePassword_FullMethodName  = "/api.Dex/CreatePassword"
+	Dex_UpdatePassword_FullMethodName  = "/api.Dex/UpdatePassword"
+	Dex_DeletePassword_FullMethodName  = "/api.Dex/DeletePassword"
+	Dex_ListPasswords_FullMethodName   = "/api.Dex/ListPasswords"
+	Dex_CreateConnector_FullMethodName = "/api.Dex/CreateConnector"
+	Dex_UpdateConnector_FullMethodName = "/api.Dex/UpdateConnector"
+	Dex_DeleteConnector_FullMethodName = "/api.Dex/DeleteConnector"
+	Dex_ListConnectors_FullMethodName  = "/api.Dex/ListConnectors"
+	Dex_GetVersion_FullMethodName      = "/api.Dex/GetVersion"
+	Dex_GetDiscovery_FullMethodName    = "/api.Dex/GetDiscovery"
+	Dex_ListRefresh_FullMethodName     = "/api.Dex/ListRefresh"
+	Dex_RevokeRefresh_FullMethodName   = "/api.Dex/RevokeRefresh"
+	Dex_VerifyPassword_FullMethodName  = "/api.Dex/VerifyPassword"
 )
 
 // DexClient is the client API for Dex service.
 //
 // For semantics around ctx use and closing/ending streaming RPCs, please refer to https://pkg.go.dev/google.golang.org/grpc/?tab=doc#ClientConn.NewStream.
+//
+// Dex represents the dex gRPC service.
 type DexClient interface {
 	// GetClient gets a client.
 	GetClient(ctx context.Context, in *GetClientReq, opts ...grpc.CallOption) (*GetClientResp, error)
@@ -53,8 +60,18 @@ type DexClient interface {
 	DeletePassword(ctx context.Context, in *DeletePasswordReq, opts ...grpc.CallOption) (*DeletePasswordResp, error)
 	// ListPassword lists all password entries.
 	ListPasswords(ctx context.Context, in *ListPasswordReq, opts ...grpc.CallOption) (*ListPasswordResp, error)
+	// CreateConnector creates a connector.
+	CreateConnector(ctx context.Context, in *CreateConnectorReq, opts ...grpc.CallOption) (*CreateConnectorResp, error)
+	// UpdateConnector modifies existing connector.
+	UpdateConnector(ctx context.Context, in *UpdateConnectorReq, opts ...grpc.CallOption) (*UpdateConnectorResp, error)
+	// DeleteConnector deletes the connector.
+	DeleteConnector(ctx context.Context, in *DeleteConnectorReq, opts ...grpc.CallOption) (*DeleteConnectorResp, error)
+	// ListConnectors lists all connector entries.
+	ListConnectors(ctx context.Context, in *ListConnectorReq, opts ...grpc.CallOption) (*ListConnectorResp, error)
 	// GetVersion returns version information of the server.
 	GetVersion(ctx context.Context, in *VersionReq, opts ...grpc.CallOption) (*VersionResp, error)
+	// GetDiscovery returns discovery information of the server.
+	GetDiscovery(ctx context.Context, in *DiscoveryReq, opts ...grpc.CallOption) (*DiscoveryResp, error)
 	// ListRefresh lists all the refresh token entries for a particular user.
 	ListRefresh(ctx context.Context, in *ListRefreshReq, opts ...grpc.CallOption) (*ListRefreshResp, error)
 	// RevokeRefresh revokes the refresh token for the provided user-client pair.
@@ -74,8 +91,9 @@ func NewDexClient(cc grpc.ClientConnInterface) DexClient {
 }
 
 func (c *dexClient) GetClient(ctx context.Context, in *GetClientReq, opts ...grpc.CallOption) (*GetClientResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(GetClientResp)
-	err := c.cc.Invoke(ctx, Dex_GetClient_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Dex_GetClient_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -83,8 +101,9 @@ func (c *dexClient) GetClient(ctx context.Context, in *GetClientReq, opts ...grp
 }
 
 func (c *dexClient) CreateClient(ctx context.Context, in *CreateClientReq, opts ...grpc.CallOption) (*CreateClientResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CreateClientResp)
-	err := c.cc.Invoke(ctx, Dex_CreateClient_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Dex_CreateClient_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -92,8 +111,9 @@ func (c *dexClient) CreateClient(ctx context.Context, in *CreateClientReq, opts 
 }
 
 func (c *dexClient) UpdateClient(ctx context.Context, in *UpdateClientReq, opts ...grpc.CallOption) (*UpdateClientResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(UpdateClientResp)
-	err := c.cc.Invoke(ctx, Dex_UpdateClient_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Dex_UpdateClient_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -101,8 +121,9 @@ func (c *dexClient) UpdateClient(ctx context.Context, in *UpdateClientReq, opts 
 }
 
 func (c *dexClient) DeleteClient(ctx context.Context, in *DeleteClientReq, opts ...grpc.CallOption) (*DeleteClientResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(DeleteClientResp)
-	err := c.cc.Invoke(ctx, Dex_DeleteClient_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Dex_DeleteClient_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -110,8 +131,9 @@ func (c *dexClient) DeleteClient(ctx context.Context, in *DeleteClientReq, opts 
 }
 
 func (c *dexClient) CreatePassword(ctx context.Context, in *CreatePasswordReq, opts ...grpc.CallOption) (*CreatePasswordResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(CreatePasswordResp)
-	err := c.cc.Invoke(ctx, Dex_CreatePassword_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Dex_CreatePassword_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -119,8 +141,9 @@ func (c *dexClient) CreatePassword(ctx context.Context, in *CreatePasswordReq, o
 }
 
 func (c *dexClient) UpdatePassword(ctx context.Context, in *UpdatePasswordReq, opts ...grpc.CallOption) (*UpdatePasswordResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(UpdatePasswordResp)
-	err := c.cc.Invoke(ctx, Dex_UpdatePassword_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Dex_UpdatePassword_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -128,8 +151,9 @@ func (c *dexClient) UpdatePassword(ctx context.Context, in *UpdatePasswordReq, o
 }
 
 func (c *dexClient) DeletePassword(ctx context.Context, in *DeletePasswordReq, opts ...grpc.CallOption) (*DeletePasswordResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(DeletePasswordResp)
-	err := c.cc.Invoke(ctx, Dex_DeletePassword_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Dex_DeletePassword_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -137,8 +161,49 @@ func (c *dexClient) DeletePassword(ctx context.Context, in *DeletePasswordReq, o
 }
 
 func (c *dexClient) ListPasswords(ctx context.Context, in *ListPasswordReq, opts ...grpc.CallOption) (*ListPasswordResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListPasswordResp)
-	err := c.cc.Invoke(ctx, Dex_ListPasswords_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Dex_ListPasswords_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *dexClient) CreateConnector(ctx context.Context, in *CreateConnectorReq, opts ...grpc.CallOption) (*CreateConnectorResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(CreateConnectorResp)
+	err := c.cc.Invoke(ctx, Dex_CreateConnector_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *dexClient) UpdateConnector(ctx context.Context, in *UpdateConnectorReq, opts ...grpc.CallOption) (*UpdateConnectorResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(UpdateConnectorResp)
+	err := c.cc.Invoke(ctx, Dex_UpdateConnector_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *dexClient) DeleteConnector(ctx context.Context, in *DeleteConnectorReq, opts ...grpc.CallOption) (*DeleteConnectorResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(DeleteConnectorResp)
+	err := c.cc.Invoke(ctx, Dex_DeleteConnector_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *dexClient) ListConnectors(ctx context.Context, in *ListConnectorReq, opts ...grpc.CallOption) (*ListConnectorResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(ListConnectorResp)
+	err := c.cc.Invoke(ctx, Dex_ListConnectors_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -146,8 +211,19 @@ func (c *dexClient) ListPasswords(ctx context.Context, in *ListPasswordReq, opts
 }
 
 func (c *dexClient) GetVersion(ctx context.Context, in *VersionReq, opts ...grpc.CallOption) (*VersionResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(VersionResp)
-	err := c.cc.Invoke(ctx, Dex_GetVersion_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Dex_GetVersion_FullMethodName, in, out, cOpts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
+func (c *dexClient) GetDiscovery(ctx context.Context, in *DiscoveryReq, opts ...grpc.CallOption) (*DiscoveryResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
+	out := new(DiscoveryResp)
+	err := c.cc.Invoke(ctx, Dex_GetDiscovery_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -155,8 +231,9 @@ func (c *dexClient) GetVersion(ctx context.Context, in *VersionReq, opts ...grpc
 }
 
 func (c *dexClient) ListRefresh(ctx context.Context, in *ListRefreshReq, opts ...grpc.CallOption) (*ListRefreshResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(ListRefreshResp)
-	err := c.cc.Invoke(ctx, Dex_ListRefresh_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Dex_ListRefresh_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -164,8 +241,9 @@ func (c *dexClient) ListRefresh(ctx context.Context, in *ListRefreshReq, opts ..
 }
 
 func (c *dexClient) RevokeRefresh(ctx context.Context, in *RevokeRefreshReq, opts ...grpc.CallOption) (*RevokeRefreshResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(RevokeRefreshResp)
-	err := c.cc.Invoke(ctx, Dex_RevokeRefresh_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Dex_RevokeRefresh_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -173,8 +251,9 @@ func (c *dexClient) RevokeRefresh(ctx context.Context, in *RevokeRefreshReq, opt
 }
 
 func (c *dexClient) VerifyPassword(ctx context.Context, in *VerifyPasswordReq, opts ...grpc.CallOption) (*VerifyPasswordResp, error) {
+	cOpts := append([]grpc.CallOption{grpc.StaticMethod()}, opts...)
 	out := new(VerifyPasswordResp)
-	err := c.cc.Invoke(ctx, Dex_VerifyPassword_FullMethodName, in, out, opts...)
+	err := c.cc.Invoke(ctx, Dex_VerifyPassword_FullMethodName, in, out, cOpts...)
 	if err != nil {
 		return nil, err
 	}
@@ -183,7 +262,9 @@ func (c *dexClient) VerifyPassword(ctx context.Context, in *VerifyPasswordReq, o
 
 // DexServer is the server API for Dex service.
 // All implementations must embed UnimplementedDexServer
-// for forward compatibility
+// for forward compatibility.
+//
+// Dex represents the dex gRPC service.
 type DexServer interface {
 	// GetClient gets a client.
 	GetClient(context.Context, *GetClientReq) (*GetClientResp, error)
@@ -201,8 +282,18 @@ type DexServer interface {
 	DeletePassword(context.Context, *DeletePasswordReq) (*DeletePasswordResp, error)
 	// ListPassword lists all password entries.
 	ListPasswords(context.Context, *ListPasswordReq) (*ListPasswordResp, error)
+	// CreateConnector creates a connector.
+	CreateConnector(context.Context, *CreateConnectorReq) (*CreateConnectorResp, error)
+	// UpdateConnector modifies existing connector.
+	UpdateConnector(context.Context, *UpdateConnectorReq) (*UpdateConnectorResp, error)
+	// DeleteConnector deletes the connector.
+	DeleteConnector(context.Context, *DeleteConnectorReq) (*DeleteConnectorResp, error)
+	// ListConnectors lists all connector entries.
+	ListConnectors(context.Context, *ListConnectorReq) (*ListConnectorResp, error)
 	// GetVersion returns version information of the server.
 	GetVersion(context.Context, *VersionReq) (*VersionResp, error)
+	// GetDiscovery returns discovery information of the server.
+	GetDiscovery(context.Context, *DiscoveryReq) (*DiscoveryResp, error)
 	// ListRefresh lists all the refresh token entries for a particular user.
 	ListRefresh(context.Context, *ListRefreshReq) (*ListRefreshResp, error)
 	// RevokeRefresh revokes the refresh token for the provided user-client pair.
@@ -214,9 +305,12 @@ type DexServer interface {
 	mustEmbedUnimplementedDexServer()
 }
 
-// UnimplementedDexServer must be embedded to have forward compatible implementations.
-type UnimplementedDexServer struct {
-}
+// UnimplementedDexServer must be embedded to have
+// forward compatible implementations.
+//
+// NOTE: this should be embedded by value instead of pointer to avoid a nil
+// pointer dereference when methods are called.
+type UnimplementedDexServer struct{}
 
 func (UnimplementedDexServer) GetClient(context.Context, *GetClientReq) (*GetClientResp, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetClient not implemented")
@@ -242,8 +336,23 @@ func (UnimplementedDexServer) DeletePassword(context.Context, *DeletePasswordReq
 func (UnimplementedDexServer) ListPasswords(context.Context, *ListPasswordReq) (*ListPasswordResp, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListPasswords not implemented")
 }
+func (UnimplementedDexServer) CreateConnector(context.Context, *CreateConnectorReq) (*CreateConnectorResp, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method CreateConnector not implemented")
+}
+func (UnimplementedDexServer) UpdateConnector(context.Context, *UpdateConnectorReq) (*UpdateConnectorResp, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method UpdateConnector not implemented")
+}
+func (UnimplementedDexServer) DeleteConnector(context.Context, *DeleteConnectorReq) (*DeleteConnectorResp, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method DeleteConnector not implemented")
+}
+func (UnimplementedDexServer) ListConnectors(context.Context, *ListConnectorReq) (*ListConnectorResp, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method ListConnectors not implemented")
+}
 func (UnimplementedDexServer) GetVersion(context.Context, *VersionReq) (*VersionResp, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetVersion not implemented")
+}
+func (UnimplementedDexServer) GetDiscovery(context.Context, *DiscoveryReq) (*DiscoveryResp, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetDiscovery not implemented")
 }
 func (UnimplementedDexServer) ListRefresh(context.Context, *ListRefreshReq) (*ListRefreshResp, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method ListRefresh not implemented")
@@ -255,6 +364,7 @@ func (UnimplementedDexServer) VerifyPassword(context.Context, *VerifyPasswordReq
 	return nil, status.Errorf(codes.Unimplemented, "method VerifyPassword not implemented")
 }
 func (UnimplementedDexServer) mustEmbedUnimplementedDexServer() {}
+func (UnimplementedDexServer) testEmbeddedByValue()             {}
 
 // UnsafeDexServer may be embedded to opt out of forward compatibility for this service.
 // Use of this interface is not recommended, as added methods to DexServer will
@@ -264,6 +374,13 @@ type UnsafeDexServer interface {
 }
 
 func RegisterDexServer(s grpc.ServiceRegistrar, srv DexServer) {
+	// If the following call pancis, it indicates UnimplementedDexServer was
+	// embedded by pointer and is nil.  This will cause panics if an
+	// unimplemented method is ever invoked, so we test this at initialization
+	// time to prevent it from happening at runtime later due to I/O.
+	if t, ok := srv.(interface{ testEmbeddedByValue() }); ok {
+		t.testEmbeddedByValue()
+	}
 	s.RegisterService(&Dex_ServiceDesc, srv)
 }
 
@@ -411,6 +528,78 @@ func _Dex_ListPasswords_Handler(srv interface{}, ctx context.Context, dec func(i
 	return interceptor(ctx, in, info, handler)
 }
 
+func _Dex_CreateConnector_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(CreateConnectorReq)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(DexServer).CreateConnector(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: Dex_CreateConnector_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(DexServer).CreateConnector(ctx, req.(*CreateConnectorReq))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _Dex_UpdateConnector_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(UpdateConnectorReq)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(DexServer).UpdateConnector(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: Dex_UpdateConnector_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(DexServer).UpdateConnector(ctx, req.(*UpdateConnectorReq))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _Dex_DeleteConnector_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(DeleteConnectorReq)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(DexServer).DeleteConnector(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: Dex_DeleteConnector_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(DexServer).DeleteConnector(ctx, req.(*DeleteConnectorReq))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _Dex_ListConnectors_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(ListConnectorReq)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(DexServer).ListConnectors(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: Dex_ListConnectors_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(DexServer).ListConnectors(ctx, req.(*ListConnectorReq))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 func _Dex_GetVersion_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
 	in := new(VersionReq)
 	if err := dec(in); err != nil {
@@ -425,6 +614,24 @@ func _Dex_GetVersion_Handler(srv interface{}, ctx context.Context, dec func(inte
 	}
 	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
 		return srv.(DexServer).GetVersion(ctx, req.(*VersionReq))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
+func _Dex_GetDiscovery_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(DiscoveryReq)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(DexServer).GetDiscovery(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: Dex_GetDiscovery_FullMethodName,
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(DexServer).GetDiscovery(ctx, req.(*DiscoveryReq))
 	}
 	return interceptor(ctx, in, info, handler)
 }
@@ -523,8 +730,28 @@ var Dex_ServiceDesc = grpc.ServiceDesc{
 			Handler:    _Dex_ListPasswords_Handler,
 		},
 		{
+			MethodName: "CreateConnector",
+			Handler:    _Dex_CreateConnector_Handler,
+		},
+		{
+			MethodName: "UpdateConnector",
+			Handler:    _Dex_UpdateConnector_Handler,
+		},
+		{
+			MethodName: "DeleteConnector",
+			Handler:    _Dex_DeleteConnector_Handler,
+		},
+		{
+			MethodName: "ListConnectors",
+			Handler:    _Dex_ListConnectors_Handler,
+		},
+		{
 			MethodName: "GetVersion",
 			Handler:    _Dex_GetVersion_Handler,
+		},
+		{
+			MethodName: "GetDiscovery",
+			Handler:    _Dex_GetDiscovery_Handler,
 		},
 		{
 			MethodName: "ListRefresh",
