@@ -7,6 +7,7 @@ import (
 	"fmt"
 	"math"
 
+	"entgo.io/ent"
 	"entgo.io/ent/dialect/sql"
 	"entgo.io/ent/dialect/sql/sqlgraph"
 	"entgo.io/ent/schema/field"
@@ -60,7 +61,7 @@ func (kq *KeysQuery) Order(o ...keys.OrderOption) *KeysQuery {
 // First returns the first Keys entity from the query.
 // Returns a *NotFoundError when no Keys was found.
 func (kq *KeysQuery) First(ctx context.Context) (*Keys, error) {
-	nodes, err := kq.Limit(1).All(setContextOp(ctx, kq.ctx, "First"))
+	nodes, err := kq.Limit(1).All(setContextOp(ctx, kq.ctx, ent.OpQueryFirst))
 	if err != nil {
 		return nil, err
 	}
@@ -83,7 +84,7 @@ func (kq *KeysQuery) FirstX(ctx context.Context) *Keys {
 // Returns a *NotFoundError when no Keys ID was found.
 func (kq *KeysQuery) FirstID(ctx context.Context) (id string, err error) {
 	var ids []string
-	if ids, err = kq.Limit(1).IDs(setContextOp(ctx, kq.ctx, "FirstID")); err != nil {
+	if ids, err = kq.Limit(1).IDs(setContextOp(ctx, kq.ctx, ent.OpQueryFirstID)); err != nil {
 		return
 	}
 	if len(ids) == 0 {
@@ -106,7 +107,7 @@ func (kq *KeysQuery) FirstIDX(ctx context.Context) string {
 // Returns a *NotSingularError when more than one Keys entity is found.
 // Returns a *NotFoundError when no Keys entities are found.
 func (kq *KeysQuery) Only(ctx context.Context) (*Keys, error) {
-	nodes, err := kq.Limit(2).All(setContextOp(ctx, kq.ctx, "Only"))
+	nodes, err := kq.Limit(2).All(setContextOp(ctx, kq.ctx, ent.OpQueryOnly))
 	if err != nil {
 		return nil, err
 	}
@@ -134,7 +135,7 @@ func (kq *KeysQuery) OnlyX(ctx context.Context) *Keys {
 // Returns a *NotFoundError when no entities are found.
 func (kq *KeysQuery) OnlyID(ctx context.Context) (id string, err error) {
 	var ids []string
-	if ids, err = kq.Limit(2).IDs(setContextOp(ctx, kq.ctx, "OnlyID")); err != nil {
+	if ids, err = kq.Limit(2).IDs(setContextOp(ctx, kq.ctx, ent.OpQueryOnlyID)); err != nil {
 		return
 	}
 	switch len(ids) {
@@ -159,7 +160,7 @@ func (kq *KeysQuery) OnlyIDX(ctx context.Context) string {
 
 // All executes the query and returns a list of KeysSlice.
 func (kq *KeysQuery) All(ctx context.Context) ([]*Keys, error) {
-	ctx = setContextOp(ctx, kq.ctx, "All")
+	ctx = setContextOp(ctx, kq.ctx, ent.OpQueryAll)
 	if err := kq.prepareQuery(ctx); err != nil {
 		return nil, err
 	}
@@ -181,7 +182,7 @@ func (kq *KeysQuery) IDs(ctx context.Context) (ids []string, err error) {
 	if kq.ctx.Unique == nil && kq.path != nil {
 		kq.Unique(true)
 	}
-	ctx = setContextOp(ctx, kq.ctx, "IDs")
+	ctx = setContextOp(ctx, kq.ctx, ent.OpQueryIDs)
 	if err = kq.Select(keys.FieldID).Scan(ctx, &ids); err != nil {
 		return nil, err
 	}
@@ -199,7 +200,7 @@ func (kq *KeysQuery) IDsX(ctx context.Context) []string {
 
 // Count returns the count of the given query.
 func (kq *KeysQuery) Count(ctx context.Context) (int, error) {
-	ctx = setContextOp(ctx, kq.ctx, "Count")
+	ctx = setContextOp(ctx, kq.ctx, ent.OpQueryCount)
 	if err := kq.prepareQuery(ctx); err != nil {
 		return 0, err
 	}
@@ -217,7 +218,7 @@ func (kq *KeysQuery) CountX(ctx context.Context) int {
 
 // Exist returns true if the query has elements in the graph.
 func (kq *KeysQuery) Exist(ctx context.Context) (bool, error) {
-	ctx = setContextOp(ctx, kq.ctx, "Exist")
+	ctx = setContextOp(ctx, kq.ctx, ent.OpQueryExist)
 	switch _, err := kq.FirstID(ctx); {
 	case IsNotFound(err):
 		return false, nil
@@ -449,7 +450,7 @@ func (kgb *KeysGroupBy) Aggregate(fns ...AggregateFunc) *KeysGroupBy {
 
 // Scan applies the selector query and scans the result into the given value.
 func (kgb *KeysGroupBy) Scan(ctx context.Context, v any) error {
-	ctx = setContextOp(ctx, kgb.build.ctx, "GroupBy")
+	ctx = setContextOp(ctx, kgb.build.ctx, ent.OpQueryGroupBy)
 	if err := kgb.build.prepareQuery(ctx); err != nil {
 		return err
 	}
@@ -497,7 +498,7 @@ func (ks *KeysSelect) Aggregate(fns ...AggregateFunc) *KeysSelect {
 
 // Scan applies the selector query and scans the result into the given value.
 func (ks *KeysSelect) Scan(ctx context.Context, v any) error {
-	ctx = setContextOp(ctx, ks.ctx, "Select")
+	ctx = setContextOp(ctx, ks.ctx, ent.OpQuerySelect)
 	if err := ks.prepareQuery(ctx); err != nil {
 		return err
 	}
