@@ -1,3 +1,3 @@
 package ent
 
-//go:generate go run -mod=mod entgo.io/ent/cmd/ent generate ./schema --target ./db
+//go:generate go tool entgo.io/ent/cmd/ent generate ./schema --target ./db
